requirements.txt
asgiref==3.6.0
async-timeout==4.0.2
cffi==1.15.1
cryptography==40.0.2
Django==4.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
Pillow==9.5.0
pycparser==2.21
pycryptodomex==3.17
PyJWT==2.6.0
pyOpenSSL==23.1.1
pytz==2023.3
redis==4.5.4
sqlparse==0.4.4
